+++ crates/models/deploy/legacy_format_instruction_data.rs
@@ -232,6 +232,7 @@ pub struct DepositData {
     pub amount: i64,
     pub lut_slot: u32,
     pub ref_id: u32,
+    pub custom_id: i64,
 }
 
 #[repr(C)]
@@ -344,6 +345,7 @@ pub struct WithdrawData {
     padding_u16: u16,
     pub token_id: u32,
     pub amount: i64,
+    pub custom_id: i64,
 }
 
 #[repr(C)]