+++ crates/models/src/log.rs
@@ -94,6 +94,7 @@ pub struct DepositReport {
     pub token_id: u32,
     pub time: u32,
     pub amount: i64,
+    pub custom_id: i64,
 }
 
 #[repr(C)]
@@ -167,6 +168,7 @@ pub struct WithdrawReport {
     pub token_id: u32,
     pub time: u32,
     pub amount: i64,
+    pub custom_id: i64,
 }
 
 #[repr(C)]