+++ crates/models/src/instruction_data.rs
@@ -397,9 +397,13 @@ pub struct SwapData {
 /// - `instr_id` - Instr pair id
 pub struct SpotQuotesReplaceData {
     pub tag: u8,
-    pub padding_u8: u8, // <- rename to bump for bots
+    pub bump: u8,
+    pub order_type: u8,
+    pub padding_u8: u8,
     pub mask: QuoteMask,
+    pub padding_u16: u16,
     pub instr_id: InstrId,
+    pub padding_u32: u32,
 }
 
 #[repr(C)]
@@ -415,9 +419,13 @@ pub struct SpotQuotesReplaceData {
 /// - `instr_id` - Instr pair id
 pub struct PerpQuotesReplaceData {
     pub tag: u8,
-    padding_u8: u8, // <- rename to bump for bots
+    pub bump: u8,
+    pub order_type: u8,
+    pub padding_u8: u8,
     pub mask: QuoteMask,
+    pub padding_u16: u16,
     pub instr_id: InstrId,
+    pub padding_u32: u32,
 }
 
 #[repr(C)]